Create DAY 97-Delete From End
OUTPUT-
Original List: 
1 2 3 4 
Updated List: 
1 2 3 
Updated List: 
1 2 
Updated List: 
1 
Updated List: 
List is empty

In this program, we will create a singly linked list and delete a node from the end of the list. To accomplish this task, we first find out the second last node of the list. Then, make the second last node as the new tail of the list. Then, delete the last node of the list.

Algorithm
1.Create a class Node which has two attributes: data and next. Next is a pointer to the next node in the list.
2.Create another class DeleteEnd which has two attributes: head and tail.
addNode() will add a new node to the list:
3.Create a new node.
4.It first checks, whether the head is equal to null which means the list is empty.
5.If the list is empty, both head and tail will point to a newly added node.
6.If the list is not empty, the new node will be added to end of the list such that tail's next will point to a newly added node. This new node will become the new tail of the list.

a. DeleteFromEnd() will delete a node from the end of the list:
1.It first checks whether the head is null (empty list) then, display the message "List is empty" and return.
2.If the list is not empty, it will check whether the list has only one node.
3.If the list has only one node, it will set both head and tail to null.
4.If the list has more than one node then, traverse through the list till node current points to second last node in the list.
5.Node current will become the new tail of the list.
6.Node next to current will be made null to delete the last node.

a. display() will display the nodes present in the list:
1.Define a node current which will initially point to the head of the list.
2.Traverse through the list till current points to null.
3.Display each node by making current to point to node next to it in each iteration.