Create DAY 98-Delete From Middle
OUTPUT-
Original List: 
1 2 3 4 
Updated List: 
1 3 4 
Updated List: 
1 4 
Updated List: 
4 
Updated List: 
List is empty

ABOUT-
In this program, we will create a singly linked list and delete a node from the middle of the list. To accomplish this task, we will calculate the size of the list and then divide it by 2 to get the mid-point of the list. Node temp will point to head node. We will iterate through the list till midpoint is reached. Now, the temp will point to middle node and node current will point to node previous to temp. We delete the middle node such that current's next node will point to temp's next node.

Algorithm
1.Create a class Node which has two attributes: data and next. Next is a pointer to the next node in the list.
2.Create another class deleteMid which has three attributes: head, tail, and size which keep tracks of the number of nodes present in the list.
addNode() will add a new node to the list:
3.Create a new node.
4.It first checks, whether the head is equal to null which means the list is empty.
5.If the list is empty, both head and tail will point to the newly added node.
6.If the list is not empty, the new node will be added to end of the list such that tail's next will point to the newly added node. This new node will become the new tail of the list.

a. deleteFromMid() will delete a node from the middle of the list:
1.It first checks whether the head is null (empty list) then, display the message "List is empty" and return.
2.If the list is not empty, it will check whether the list has only one node.
3.If the list has only one node, it will set both head and tail to null.
4.If the list has more than one node then, calculate the size of the list and divide it by 2 to get the mid-point of the list.
5.Declare a node temp which will point to head and node current will point to node previous to temp.
6.Traverse through the list till temp points to a middle node. If current not point to null then, delete the middle node(temp) by making current's next to point to temp's next. Else, both head and tail will point to node next to temp and delete the middle node by setting the temp to null.

a. display() will display the nodes present in the list:
1.Define a node current which will initially point to the head of the list.
2.Traverse through the list till current points to null.
3.Display each node by making current to point to node next to it in each iteration.